refactor: News section
